refactor(metasrv): simplify event module names
