+++ src/meta-srv/src/event.rs
@@ -20,8 +20,8 @@ use common_event_recorder::error::{InsertEventsSnafu, Result};
 use common_event_recorder::{Event, EventHandler, build_row_inserts_request, group_events_by_type};
 use snafu::ResultExt;
 
-pub mod region_migration_event;
-pub mod repartition_event;
+pub mod region_migration;
+pub mod repartition;
 
 /// EventHandlerImpl is the default event handler implementation in metasrv.
 /// It sends the received events to the frontend instances.
